+++ Channel EasyFrontEnd-HoiDanIT-TrungQuanDev/HoiDanIT/Redux Ultimate/src/App.jsx
@@ -1,10 +1,6 @@
 import './App.css'
-import { connect } from 'react-redux'
 import { increaseCounter, decreaseCounter } from './action/actions'
-import store from './redux/store'
 import { useSelector, useDispatch } from 'react-redux'
-import axios from 'axios'
-import { useEffect } from 'react'
 import Home from './components/Home'
 
 const App = (props) => {
@@ -21,16 +17,6 @@ const App = (props) => {
     // bên trong dispatch là 1 action -> action là 1 function
     dispatch(decreaseCounter())
   }
-
-  const fetchAllUser = async () => {
-    const res = await axios.get('http://localhost:8080/users/all')
-    const data = res && res.data ? res.data : []
-    console.log('>>> check data axios:', data)
-  }
-
-  useEffect(() => {
-    fetchAllUser()
-  }, [])
 
   return (
     <div className='App'>
